% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries_qaqc.R
\name{plot_timeseries_qaqc}
\alias{plot_timeseries_qaqc}
\title{Plot a time series of QA/QC'ed data}
\usage{
plot_timeseries_qaqc(
  qaqced_ts_data,
  date_col = NULL,
  value_cols = NULL,
  time_step = NA,
  flag_prefix = ".flag_",
  log_scale_time_series = FALSE,
  log_scale_density = FALSE
)
}
\arguments{
\item{qaqced_ts_data}{data frame containing QA/QC'ed data - typically the output of \code{\link[=qaqc_timeseries]{qaqc_timeseries()}}}

\item{date_col}{A tidyselect expression that resolves to a single column in \code{ts_data} with dates.
Default is \code{NULL}, which will use the first date column from \code{ts_data}.}

\item{value_cols}{A tidyselect expression that resolves to one or more columns in \code{ts_data} with values to check.
Default is \code{NULL}, which will use \code{dplyr::where(is.numeric)} to get all numeric columns from \code{ts_data}.}

\item{time_step}{A character string or a \code{Period} object, the time
step of the timeseries. If \code{NA} provided, the most common time step is determined using \link[handyr:get_step]{handyr::get_step}.}

\item{flag_prefix}{character string specifying the prefix for the flag columns}

\item{log_scale_time_series}{logical indicating whether to use log scale for the time series}

\item{log_scale_density}{logical indicating whether to use log scale for the density plot}
}
\description{
Plot a time series of QA/QC'ed data
}
