% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc_timeseries.R
\name{qaqc_timeseries}
\alias{qaqc_timeseries}
\title{Check for issues in timeseries data.}
\usage{
qaqc_timeseries(
  ts_data,
  date_col,
  value_cols,
  time_step = NA,
  precision = Inf,
  rolling = NA,
  allowed_range = c(-Inf, Inf),
  allowed_steps = list(`1 hours` = Inf),
  allowed_repeats = Inf
)
}
\arguments{
\item{ts_data}{A data frame with at least a date column and one or more value
columns.}

\item{date_col}{A tidyselect expression that resolves to a single column in \code{ts_data} with dates.}

\item{value_cols}{A tidyselect expression that resolves to one or more columns in \code{ts_data} with values to check.}

\item{time_step}{A character string or a \code{Period} object, the time
step of the timeseries. If \code{NA} provided, the most common time step is determined using \link[handyr:get_step]{handyr::get_step}.}

\item{precision}{A numeric value, the precision to which \code{value_cols} are
rounded before assessments. Default is \code{Inf}, meaning no rounding is done.}

\item{rolling}{A character string or a \code{Period} object (such as \code{"3 hours"}), the time step of observations to average over for rolling average checks.
Default is \code{NA}, meaning no rolling average checking is done.}

\item{allowed_range}{A numeric vector of length 2, the allowed range of
values. Default is \code{c(-Inf, Inf)}, meaning no range checking is done.}

\item{allowed_steps}{A named list, the allowed steps in the timeseries.
Each element of the list is a named numeric vector of length 1, the
name is the time step to look over and the value is the maximum allowed change within that time step.
Default is \code{list("1 hours" = Inf)}, meaning no step checking is done.}

\item{allowed_repeats}{A numeric value, the maximum allowed number of
consecutive repeated values. Default is \code{Inf}, meaning no repeat checking is done.}
}
\value{
A data frame with the same columns as \code{ts_data}, plus two columns
for each value column, the first containing a binary number combining the raised flags,
and the second containing a list of the individual raised flags.
}
\description{
Check for issues in timeseries data.
}
